requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.3
django-extensions==3.2.3
django-ninja==1.3.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-apis-oauth-django==1.0.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
MarkupSafe==3.0.2
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.6.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
